+++ R/GWAS.R
@@ -1,37 +1,6 @@
-#' Performs Single Marker Regressions Using BGData Objects.
-#'
-#' Implements single marker regressions. The regression model includes all the
-#' covariates specified in the right-hand-side of the `formula` plus one column
-#' of `@@geno` at a time. The data from the association tests is obtained from
-#' a [BGData-class] object.
-#'
-#' @inheritSection BGData-package File-backed matrices
-#' @inheritSection BGData-package Multi-level parallelism
-#' @param formula The formula for the GWAS model without including the marker,
-#' e.g. `y ~ 1` or `y ~ factor(sex) + age`. The variables included in the
-#' formula must be in the `@@pheno` object of the [BGData-class].
-#' @param data A [BGData-class] object.
-#' @param method The regression method to be used. Currently, the following
-#' methods are implemented: `rayOLS`, [stats::lsfit()], [stats::lm()],
-#' [stats::lm.fit()], [stats::glm()], [lme4::lmer()], and [SKAT::SKAT()].
-#' Defaults to `lsfit`.
-#' @param i Indicates which rows of `@@geno` should be used. Can be integer,
-#' boolean, or character. By default, all rows are used.
-#' @param j Indicates which columns of `@@geno` should be used. Can be integer,
-#' boolean, or character. By default, all columns are used.
-#' @param chunkSize The number of columns of `@@geno` that are brought into
-#' physical memory for processing per core. If `NULL`, all elements in `j` are
-#' used. Defaults to 5000.
-#' @param nCores The number of cores (passed to [parallel::mclapply()]).
-#' Defaults to the number of cores as detected by [parallel::detectCores()].
-#' @param verbose Whether progress updates will be posted. Defaults to `FALSE`.
-#' @param ... Additional arguments for chunkedApply and regression method.
-#' @return The same matrix that would be returned by `coef(summary(model))`.
-#' @example man/examples/GWAS.R
-#' @export
-GWAS <- function(formula, data, method = "lsfit", i = seq_len(nrow(data@geno)), j = seq_len(ncol(data@geno)), chunkSize = 5000L, nCores = getOption("mc.cores", 2L), verbose = FALSE, ...) {
+GWAS <- function(formula, data, method = "lsfit", i = seq_len(nrow(geno(data))), j = seq_len(ncol(geno(data))), chunkSize = 5000L, nCores = getOption("mc.cores", 2L), verbose = FALSE, ...) {
 
-    if (class(data) != "BGData") {
+    if (!inherits(data, "BGData")) {
         stop("data must BGData")
     }
 
@@ -39,11 +8,11 @@ GWAS <- function(formula, data, method = "lsfit", i = seq_len(nrow(data@geno)), 
         stop("Only rayOLS, lsfit, lm, lm.fit, glm, lmer, and SKAT have been implemented so far.")
     }
 
-    i <- crochet::convertIndex(data@geno, i, "i")
-    j <- crochet::convertIndex(data@geno, j, "j")
+    i <- convertIndex(geno(data), i, "i")
+    j <- convertIndex(geno(data), j, "j")
 
     if (method == "rayOLS") {
-        if (length(labels(stats::terms(formula))) > 0L) {
+        if (length(labels(terms(formula))) > 0L) {
             stop("method rayOLS can only be used with y~1 formula, if you want to add covariates pre-adjust your phenotype.")
         }
         OUT <- GWAS.rayOLS(formula = formula, data = data, i = i, j = j, chunkSize = chunkSize, nCores = nCores, verbose = verbose, ...)
@@ -63,59 +32,56 @@ GWAS <- function(formula, data, method = "lsfit", i = seq_len(nrow(data@geno)), 
         } else {
             FUN <- match.fun(method)
         }
-        GWAS.model <- stats::update(formula, ".~z+.")
-        OUT <- chunkedApply(X = data@geno, MARGIN = 2L, FUN = function(col, ...) {
-            df <- data@pheno[i, , drop = FALSE]
-            df$z <- col
+        GWAS.model <- update(formula, ".~z+.")
+        OUT <- chunkedApply(X = geno(data), MARGIN = 2L, FUN = function(col, ...) {
+            df <- pheno(data)[i, , drop = FALSE]
+            df[["z"]] <- col
             fm <- FUN(GWAS.model, data = df, ...)
             getCoefficients(fm)
         }, i = i, j = j, chunkSize = chunkSize, nCores = nCores, verbose = verbose, ...)
         OUT <- t(OUT)
-        rownames(OUT) <- colnames(data@geno)[j]
+        rownames(OUT) <- colnames(geno(data))[j]
     }
 
     return(OUT)
 }
 
-
-GWAS.rayOLS <- function(formula, data, i = seq_len(nrow(data@geno)), j = seq_len(ncol(data@geno)), chunkSize = 5000L, nCores = getOption("mc.cores", 2L), verbose = FALSE, ...) {
-    y <- data@pheno[i, getResponse(formula)]
+GWAS.rayOLS <- function(formula, data, i = seq_len(nrow(geno(data))), j = seq_len(ncol(geno(data))), chunkSize = 5000L, nCores = getOption("mc.cores", 2L), verbose = FALSE, ...) {
+    y <- pheno(data)[i, getResponse(formula)]
     y <- as.numeric(y)
-    res <- chunkedMap(X = data@geno, FUN = rayOLS, i = i, j = j, chunkSize = chunkSize, nCores = nCores, verbose = verbose, y = y, ...)
+    res <- chunkedMap(X = geno(data), FUN = rayOLS, i = i, j = j, chunkSize = chunkSize, nCores = nCores, verbose = verbose, y = y, ...)
     res <- do.call(rbind, res)
-    colnames(res) <- c("Estimate", "Std.Err", "t-value", "Pr(>|t|)")
-    rownames(res) <- colnames(data@geno)[j]
+    colnames(res) <- c("Estimate", "Std.Err", "t-value", "Pr(>|t|)", "n", "allele_freq")
+    rownames(res) <- colnames(geno(data))[j]
     return(res)
 }
 
-
-GWAS.lsfit <- function(formula, data, i = seq_len(nrow(data@geno)), j = seq_len(ncol(data@geno)), chunkSize = 5000L, nCores = getOption("mc.cores", 2L), verbose = FALSE, ...) {
+GWAS.lsfit <- function(formula, data, i = seq_len(nrow(geno(data))), j = seq_len(ncol(geno(data))), chunkSize = 5000L, nCores = getOption("mc.cores", 2L), verbose = FALSE, ...) {
 
     # The subset argument of model.frame is evaluated in the environment of the
     # formula, therefore subset after building the frame.
-    frame <- stats::model.frame(formula = formula, data = data@pheno, na.action = stats::na.pass)[i, , drop = FALSE]
-    model <- stats::model.matrix(formula, frame)
+    frame <- model.frame(formula = formula, data = pheno(data), na.action = na.pass)[i, , drop = FALSE]
+    model <- model.matrix(formula, frame)
 
-    y <- data@pheno[i, getResponse(formula)]
+    y <- pheno(data)[i, getResponse(formula)]
 
-    res <- chunkedApply(X = data@geno, MARGIN = 2L, FUN = function(col, ...) {
-        fm <- stats::lsfit(x = cbind(col, model), y = y, intercept = FALSE)
-        stats::ls.print(fm, print.it = FALSE)$coef.table[[1L]][1L, ]
+    res <- chunkedApply(X = geno(data), MARGIN = 2L, FUN = function(col, ...) {
+        fm <- lsfit(x = cbind(col, model), y = y, intercept = FALSE)
+        ls.print(fm, print.it = FALSE)[["coef.table"]][[1L]][1L, ]
     }, i = i, j = j, chunkSize = chunkSize, nCores = nCores, verbose = verbose, ...)
     res <- t(res)
-    rownames(res) <- colnames(data@geno)[j]
+    rownames(res) <- colnames(geno(data))[j]
 
     return(res)
 }
 
-
 # formula: the formula for the GWAS model without including the markers, e.g.
 # y~1 or y~factor(sex)+age
-# all the variables in the formula must be in data@pheno data (BGData)
+# all the variables in the formula must be in data@pheno (BGData)
 # containing slots @pheno and @geno
 # groups: a vector mapping markers into groups (can be integer, character or
 # factor)
-GWAS.SKAT <- function(formula, data, groups, i = seq_len(nrow(data@geno)), j = seq_len(ncol(data@geno)), verbose = FALSE, ...) {
+GWAS.SKAT <- function(formula, data, groups, i = seq_len(nrow(geno(data))), j = seq_len(ncol(geno(data))), verbose = FALSE, ...) {
 
     uniqueGroups <- unique(groups)
 
@@ -123,12 +89,12 @@ GWAS.SKAT <- function(formula, data, groups, i = seq_len(nrow(data@geno)), j = s
     colnames(OUT) <- c("nMrk", "p-value")
     rownames(OUT) <- uniqueGroups
 
-    H0 <- SKAT::SKAT_Null_Model(formula, data = data@pheno[i, , drop = FALSE], ...)
+    H0 <- SKAT::SKAT_Null_Model(formula, data = pheno(data)[i, , drop = FALSE], ...)
 
     for (group in seq_along(uniqueGroups)) {
-        Z <- data@geno[i, groups == uniqueGroups[group], drop = FALSE]
+        Z <- geno(data)[i, groups == uniqueGroups[group], drop = FALSE]
         fm <- SKAT::SKAT(Z = Z, obj = H0, ...)
-        OUT[group, ] <- c(ncol(Z), fm$p.value)
+        OUT[group, ] <- c(ncol(Z), fm[["p.value"]])
         if (verbose) {
             message("Group ", group, " of ", length(uniqueGroups), " ...")
         }
@@ -137,30 +103,25 @@ GWAS.SKAT <- function(formula, data, groups, i = seq_len(nrow(data@geno)), j = s
     return(OUT)
 }
 
-
 rayOLS <- function(x, y) {
     .Call(C_rayOLS, x, y)
 }
-
 
 getCoefficients <- function(x) {
     UseMethod("getCoefficients")
 }
 
-
 getCoefficients.lm <- function(x) {
-    summary(x)$coef[2L, ]
+    coef(summary(x))[2L, ]
 }
-
 
 getCoefficients.glm <- function(x) {
-    summary(x)$coef[2L, ]
+    coef(summary(x))[2L, ]
 }
 
-
 getCoefficients.lmerMod <- function(x) {
-    ans <- summary(x)$coef[2L, ]
-    ans <- c(ans, c(1L - stats::pnorm(ans[3L])))
+    ans <- coef(summary(x))[2L, ]
+    ans <- c(ans, c(1L - pnorm(ans[3L])))
     return(ans)
 }