+++ src/summarize.c
@@ -1,7 +1,5 @@
-#define R_NO_REMAP
+#include "summarize.h"
 
-#include <R.h>
-#include <Rinternals.h>
 #include <Rmath.h>
 
 SEXP summarize_real(SEXP X) {
@@ -19,7 +17,7 @@ SEXP summarize_real(SEXP X) {
         // Compute number of non-missing values (n), and
         // Compute column sum (xt1), and
         // Compute column sum of squares (xtx)
-        int n = 0;
+        R_xlen_t n = 0;
         double xt1 = 0;
         double xtx = 0;
         for (row_idx = 0; row_idx < nrow; row_idx++) {
@@ -69,8 +67,8 @@ SEXP summarize_integer(SEXP X) {
         // Compute number of non-missing values (n), and
         // Compute column sum (xt1), and
         // Compute column sum of squares (xtx)
-        int n = 0;
-        long int xt1 = 0;
+        R_xlen_t n = 0;
+        double xt1 = 0;
         double xtx = 0;
         for (row_idx = 0; row_idx < nrow; row_idx++) {
             int x_val = X_data[row_idx + (col_idx * nrow)];
@@ -85,10 +83,10 @@ SEXP summarize_integer(SEXP X) {
         double sd;
         if (n) {
             // Center xtx
-            xtx -= (xt1 * xt1) / (double) n;
+            xtx -= (xt1 * xt1) / n;
             // Compute summary statistics
             freq_na = (nrow - n) / (double) nrow;
-            allele_freq = xt1 / (double) n / 2;
+            allele_freq = xt1 / n / 2;
             sd = sqrt(xtx / (n - 1));
         } else {
             freq_na = 1;